[refactor] chinese/english log

+++ src/main/java/ecnu/db/analyzer/online/node/ExecutionNode.java
@@ -90,7 +90,7 @@ public class ExecutionNode {
         return outputRows;
     }
 
-    public void setOutputRows(int outputRows) {
+    public void setOutputRows(long outputRows) {
         this.outputRows = outputRows;
     }
